feat: [new-card-bento-card-cleanup-003] - [Keep the cleaned-up new-card experience compact, accessible, and durable]
