+++ ui/scripts/dashboard-shell-storybook-responsive.test.mjs
@@ -73,6 +73,7 @@ describe("verifyDashboardShellConsolidation", () => {
     const timelineSlider = { isVisible: vi.fn().mockResolvedValue(true) };
     const timelineStatus = { isVisible: vi.fn().mockResolvedValue(true) };
     const moveButton = { isVisible: vi.fn().mockResolvedValue(true) };
+    const retiredStreamStatus = { count: vi.fn().mockResolvedValue(0) };
     const workTotalsCard = {
       evaluate: vi.fn().mockResolvedValue({ ...shellStyle }),
       isVisible: vi.fn().mockResolvedValue(true),
@@ -80,6 +81,9 @@ describe("verifyDashboardShellConsolidation", () => {
     const toolbar = {
       evaluate: vi.fn().mockResolvedValue(shellStyle),
       getByRole: vi.fn((role, options) => {
+        if (role === "status") {
+          return retiredStreamStatus;
+        }
         if (role === "slider" && options?.name === "Timeline tick") {
           return timelineSlider;
         }
